requirements.txt
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/fvcore
accelerate==0.23.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.5
backcall==0.2.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
debugpy==1.6.0
docker-pycreds==0.4.0
einops==0.6.0
executing==0.8.3
filelock==3.10.0
gitdb==4.0.10
GitPython==3.1.31
google-auth-oauthlib==0.4.6
huggingface-hub==0.13.2
hydra-core==1.3.2
idna==3.4
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
nest-asyncio==1.5.5
numpy==1.22.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==21.3
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.4.0
portalocker==2.7.0
prompt-toolkit==3.0.30
protobuf==3.20.0
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rsa==4.8
sentry-sdk==1.17.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sqlparse==0.2.2
stack-data==0.3.0
submitit==1.4.5
tabulate==0.9.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
tokenizers==0.13.2
torch-tb-profiler==0.3.1
tornado==6.2
tqdm==4.65.0
transformers==4.34.1
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
wcwidth==0.2.5
yacs==0.1.8
zipp==3.15.0
timm==0.4.12
open3d==0.17.0
plyfile==1.0.1
jsonlines==4.0.0
torch_scatter==2.1.1
albumentations==1.3.1
black==21.4b2
pydot==1.4.2
future==0.18.3
torch-scatter==2.1.1
easydict==1.10
volumentations-3D==1.0.4
pycocotools==2.0.7
ftfy==6.1.1
sentence-transformers==2.2.2
nltk==3.8.1